fix: useModel getting stuck
